x86: Use per-cpu page tables
This lays the ground for hiding truly private per-cpu data, though this
step is left for later.

CPU-local page tables allow to replace the per-cpu accessors with a
simpler mechanism: just define a region that maps the local per-cpu
structure at the same virtual address for each CPU.

As this_cpu_* now returns pointers to the local per-cpu mapping region,
we need to be careful to use it only as long as this mapping persists,
ie. from arch_cpu_init until arch_cpu_restore.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>