arm: Use per-cpu page tables
Very similar to arm64, just differing in that we switch to the CPU-local
stack mapping during EL2 enabling. Here we can simplify the existing
stack pointer calculation as we only need to move it by the delta
between physical cpu_data and LOCAL_CPU_BASE. Moreover, ID mapping
happens against the CPU-local page tables because it is done on-demand
here.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>