core: Move stack and guest registers into common per_cpu fields section
All architectures share a stack at the beginning of the per_cpu
structure. Moreover, all our archs have a stack that increases
downwards. So it's easy to share the x86 stack union with the guest
registers saved at its top. The archs just need to define STACK_SIZE as
needed.

This also allows to drop the guest_regs accessor for arm and arm64.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>