arm-common: Move GICR fields into public per-cpu section
The GICv3 fields used to access the GICR across CPUs (within cell
boundaries), thus need to be public. We also move gicc_initialized along
with them as this will simplify the shutdown by ensuring that this field
is always accessible, irrespective of the passed pointer and the
availability of the private per-cpu sections.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>